[None][perf] idea_0020_v1_perf_v1: fuse DSV4 EP localize-routing chain into one Triton kernel (idea idea_0020)
New `auto_deploy::deepseek_v4_localize_routing` Triton op collapses the per-MoE-call
chain that prepares the trtllm-gen W4A16 runner's inputs: global->local expert
subtraction, range mask, invalid-sentinel select, int32 ID cast, and routing-weight
mask/bf16-cast (lines 615-624 of `_run_trtllm_gen_mxfp4_from_routing`), plus the
upstream model-side `routing_weights.to(bf16)` cast at modeling:997. ~11 launch-bound
[num_tokens, top_k] CUDA-graph nodes per MoE call become one program emitting the exact
int32 local IDs (off-rank routes -> sentinel `local_experts` so the runner skips them)
and masked bf16 weights the runner already consumes; the runner is untouched. Bit-exact
by construction: `bf16(f32(w))==bf16(bf16(w))==bf16(w)`, so dropping the model-side cast
and casting fp32->bf16 inside the kernel is identical for every valid slot (0 otherwise).
Covers both EP and non-EP variants (single shared core). 10/10 proxy MoE layers route
through it; net -10 kernels/MoE call x 100 calls = -1000 GPU kernels.

DeepSeek-V4-Flash proxy(10L) decode window total_us (comm-excl): 61831.1 -> 60621.8 us
  (-1.96%, keep-path A removal, 6.5x instance-drop floor 0.3%);
  copy_cast instances 8750->8250 (-500), elementwise 10700->10100 (-600),
  +100 `other` (the fused kernel) => net -1000 kernels removed;
  shares copy_cast 24.83%->22.38%, elementwise 19.10%->16.76% (comm +7.25pp = NCCL
  jitter, instances 210->210 unchanged -> excluded from the gate).
Full-model tpot confirmation deferred to Tier-2 batched validation.

Accuracy (test_deepseek_v4_localize_routing): 33/33 byte-exact (torch.equal on local_idx
and weights vs the unfused chain) across decode/prefill, top_k 6/8, non-EP + 3 EP shards
with off-rank sentinel routes, and bf16-vs-fp32 input equivalence -- PASSED. Proxy numdiff
diverges only via documented random-init cross-process FP non-determinism (shared
"oraec From" generated prefix confirms no systematic error).

+++ tensorrt_llm/_torch/auto_deploy/models/custom/modeling_deepseek_v4.py
@@ -994,7 +994,10 @@ class DeepseekV4MoE(nn.Module):
         routed = torch.ops.auto_deploy.torch_mxfp4_moe_from_routing(
             hidden_states_flat,
             selected_experts,
-            routing_weights.to(hidden_states_flat.dtype),
+            # Pass fp32 routing weights straight through: the trtllm-gen path's fused
+            # ``deepseek_v4_localize_routing`` casts to bf16 internally (bit-identical to
+            # the old ``.to(bf16)`` -> f32 -> bf16 round-trip), so this cast is now dead.
+            routing_weights,
             self.experts.gate_up_proj_blocks,
             self.experts.gate_up_proj_bias,
             self.experts.gate_up_proj_scales,
